requirements.txt
asgiref==3.7.2
Django==4.2.5
Jinja2==3.1.2
MarkupSafe==2.1.3
mysqlclient==2.2.0
sqlparse==0.4.4
python-dateutil~=2.8.2